Tidying up the code
Removed old css, added a new and tidied up version (stylesheet.css). Not
done, and the rest of the code stored in progress-20140920.css

Also added a table of contents for easier navigating through the
stylesheet and new sprites to be added later on. An in depth tutorial
and a well made documentation will also be added later on.